ARP response
I got the ARP response working! This is what I learned about the
conceptual part of the program:
We were pinging the router. The gateway before our router was sending
an ARP broadcast to see if our router existed. We need to respond to
the ARP request, then the gateway will send the ping (an IP packet,
different from an ARP packet).